Restructure the repository
The nested folder structure was unnecessary.
