--- scriptrepository/README.md
@@ -1,1 +1,0 @@
-The files here support uploading scripts from Mantid through to the script repository
